pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "omix"
version = "0.1.0"
description = "A modular Python package for fetching, enriching, and analyzing omics metadata and publications."
readme = "README.md"
requires-python = ">=3.10"
license = {text = "MIT"}
authors = [{name = "Heather MacGregor", email = "user@example.com"}]
dependencies = [
    "requests>=2.31",
    "beautifulsoup4>=4.12",
    "pandas>=2.0",
    "pypdf2>=3.0",
    "urllib3>=2.0",
    "pyyaml>=6.0",
    "rich>=13.0",
    "click>=8.0",
    "aiohttp>=3.8",
    "biopython>=1.81",
    "geopy>=2.0",
    "rapidfuzz>=3.0",
    "tqdm>=4.0",
]

[project.optional-dependencies]
llm = ["openai>=1.0", "transformers>=4.30"]
dev = ["pytest", "pytest-asyncio", "httpx"]

[project.scripts]
omix = "omix.cli:main"

[tool.pytest.ini_options]
testpaths = ["tests"]
markers = [
    "integration: tests that may require network or external services",
]